Refactor: 게시글 업로드 모달 시리즈 설정 컴포넌트 분해
